LayoutTests/canvas/philip/tests/2d.drawImage.nowrap-expected.txt

< [index] >
2d.drawImage.nowrap
Stretched images do not get pixels wrapping around the edges
References:
Actual output:
Expected output:

Passed
